remove __awaiter

+++ out/replManager.js
@@ -1,12 +1,5 @@
 "use strict";
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
+
 Object.defineProperty(exports, "__esModule", { value: true });
 const vscode = require("vscode");
 const os = require('os');
@@ -17,34 +10,30 @@ class REPLManager {
     }
     //Runs the REPL using the current file.
     run(filepath) {
-        return __awaiter(this, void 0, void 0, function* () {
-            //Always run in a new terminal (I found no other way to close the Racket shell)
-            //Stop the old terminal
-            this.stop(this._terminal);
-            //Start the REPL.
-            this.launch(filepath);
-            //Focus terminal.
-            this._terminal.show(false);
-        });
+        //Always run in a new terminal (I found no other way to close the Racket shell)
+        //Stop the old terminal
+        this.stop(this._terminal);
+        //Start the REPL.
+        this.launch(filepath);
+        //Focus terminal.
+        this._terminal.show(false);
     }
     //Stops the REPL in the given terminal (defaults to running terminal).
     stop(terminal = this._terminal) {
-        return __awaiter(this, void 0, void 0, function* () {
-            terminal.hide();
-            const pid = yield terminal.processId;
-            //On windows and linux/mac require a different kill method.
-            switch (os_type) {
-                case 'win32': {
-                    const kill = require('tree-kill');
-                    kill(pid);
-                    return;
-                }
-                default: {
-                    const exec = require('child_process').exec;
-                    exec(`kill -9 ${pid}`); //kill terminal process using SIGKILL
-                }
+        if (terminal == null) { return null; }
+        const pid = terminal.processId;
+        //On windows and linux/mac require a different kill method.
+        switch (os_type) {
+            case 'win32': {
+                const kill = require('tree-kill');
+                kill(pid);
+                return;
             }
-        });
+            default: {
+                const exec = require('child_process').exec;
+                exec(`kill -9 ${pid}`); //kill terminal process using SIGKILL
+            }
+        }
     }
     //Stop REPL when object gets disposed.
     dispose() {
